# docs/swagger.yaml
basePath: /api
definitions:
...
  models.ApplicantResponse:
    properties:
      created_at:
        type: string
      date_of_birth:
        type: string
      employment_status:
        type: string
      household:
        items:
          $ref: '#/definitions/models.HouseholdMember'
        type: array
      id:
        type: string
      marital_status:
        type: string
      name:
        type: string
      sex:
        type: string
      updated_at:
        type: string
    type: object
  models.ApplicationRequest:
    properties:
      applicant_id:
        type: string
      notes:
        type: string
      scheme_id:
        type: string
    type: object
  models.Benefit:
    properties:
      amount:
        type: number
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      scheme_id:
        type: string
      updated_at:
        type: string
    type: object
  models.ChildCriteria:
    properties:
      school_level:
        type: string
    type: object
  models.Criteria:
    properties:
      employment_status:
        type: string
      has_children:
        $ref: '#/definitions/models.ChildCriteria'
      marital_status:
        type: string
    type: object
...
  models.HouseholdMember:
    properties:
      applicant_id:
        type: string
      created_at:
        type: string
      date_of_birth:
        type: string
      employment_status:
        type: string
      id:
        type: string
      name:
        type: string
      relation:
        type: string
      sex:
        type: string
      updated_at:
        type: string
    type: object
...
  models.SchemeResponse:
    properties:
      benefits:
        items:
          $ref: '#/definitions/models.Benefit'
        type: array
      created_at:
        type: string
      criteria:
        $ref: '#/definitions/models.Criteria'
      description:
        type: string
      id:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
  models.SwaggerApplicationResponse:
    description: Response containing an application with applicant and scheme details
    properties:
      applicant:
        $ref: '#/definitions/models.ApplicantResponse'
      applicant_id:
        example: 01913b7a-4493-74b2-93f8-e684c4ca935c
        type: string
      application_date:
        type: string
      created_at:
        type: string
      decision_date:
        type: string
      id:
        example: 01913b7a-4493-74b2-93f8-e684c4ca935c
        type: string
      notes:
        type: string
      scheme:
        $ref: '#/definitions/models.SchemeResponse'
      scheme_id:
        example: 01913b89-9a43-7163-8757-01cc254783f3
        type: string
      status:
        enum:
        - pending
        - approved
        - rejected
        example: pending
        type: string
      updated_at:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
paths:
...
  /api/applications:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all financial assistance applications
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.SwaggerApplicationResponse'
            type: array
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Get all applications
      tags:
      - applications
    post:
      consumes:
      - application/json
      description: Submit a new application for a financial assistance scheme
      parameters:
      - description: Application information
        in: body
        name: application
        required: true
        schema:
          $ref: '#/definitions/models.ApplicationRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.SwaggerApplicationResponse'
        "400":
          description: Bad request
          schema:
            type: string
        "404":
          description: Applicant or scheme not found
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Create a new application
      tags:
      - applications
  /api/applications/{id}:
    delete:
      consumes:
      - application/json
      description: Remove an application from the system
      parameters:
      - description: Application ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No content
        "404":
          description: Application not found
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Delete application
      tags:
      - applications
    get:
      consumes:
      - application/json
      description: Retrieve a specific application by its ID
      parameters:
      - description: Application ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SwaggerApplicationResponse'
        "404":
          description: Application not found
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Get application by ID
      tags:
      - applications
    put:
      consumes:
      - application/json
      description: Update an existing application's status or notes
      parameters:
      - description: Application ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated application information
        in: body
        name: application
        required: true
        schema:
          properties:
            notes:
              type: string
            status:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SwaggerApplicationResponse'
        "400":
          description: Bad request
          schema:
            type: string
        "404":
          description: Application not found
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Update application
      tags:
      - applications
...
schemes:
- http
swagger: "2.0"